Refactor patu async job
