+++ service/src/main/kotlin/fi/espoo/evaka/reports/patu/PatuReportingController.kt
@@ -10,14 +10,18 @@ import fi.espoo.evaka.shared.async.AsyncJobRunner
 import fi.espoo.evaka.shared.auth.AuthenticatedUser
 import fi.espoo.evaka.shared.db.Database
 import fi.espoo.evaka.shared.domain.DateRange
+import fi.espoo.evaka.shared.domain.HelsinkiDateTime
 import fi.espoo.evaka.shared.security.AccessControl
 import fi.espoo.evaka.shared.security.Action
+import mu.KotlinLogging
 import org.springframework.format.annotation.DateTimeFormat
 import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RequestParam
 import org.springframework.web.bind.annotation.RestController
 import java.time.LocalDate
+
+private val logger = KotlinLogging.logger {}
 
 @RestController
 @RequestMapping("/patu-report")
@@ -34,7 +38,14 @@ class PatuReportingController(private val asyncJobRunner: AsyncJobRunner<AsyncJo
         accessControl.requirePermissionFor(user, Action.Global.SEND_PATU_REPORT)
         db.connect { dbc ->
             dbc.transaction { tx ->
-                asyncJobRunner.plan(tx, listOf(AsyncJob.SendPatuReport(DateRange(from, to))))
+                val range = DateRange(from, to)
+                logger.info("Scheduling patu report $range")
+                asyncJobRunner.plan(
+                    tx,
+                    payloads = listOf(AsyncJob.SendPatuReport(range)),
+                    runAt = HelsinkiDateTime.now(),
+                    retryCount = 1
+                )
             }
         }
     }
